requirements.txt
python-dotenv
pydantic

openai

langchain
langchain-openai
langchain-community
langchain-text-splitters
langgraph
langsmith

faiss-cpu
tiktoken

sqlalchemy

Pillow
pdf2image
PyPDF2

streamlit
streamlit-chat

pandas